Fix circular between file & rag
