+++ src/database/server/schemas/lobechat/index.ts
@@ -3,7 +3,6 @@ export * from './discover';
 export * from './file';
 export * from './message';
 export * from './nextauth';
-export * from './rag';
 export * from './relations';
 export * from './session';
 export * from './topic';
