+++ src/database/server/schemas/lobechat/message.ts
@@ -15,8 +15,7 @@ import { createSelectSchema } from 'drizzle-zod';
 import { idGenerator } from '../../utils/idGenerator';
 import { createdAt, updatedAt } from './_helpers';
 import { agents } from './agent';
-import { files } from './file';
-import { chunks, embeddings } from './rag';
+import { chunks, embeddings, files } from './file';
 import { sessions } from './session';
 import { topics } from './topic';
 import { users } from './user';
